# packages/server/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Desk API",
    "version": "1.0.0",
    "description": "HTTP + WS gateway for the Desk personal AI assistant."
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  },
  "paths": {
...
    "/auth/login": {
      "post": {
        "summary": "Log in",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "username",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
...
    "/me": {
      "get": {
        "summary": "Get current user",
        "responses": {
          "200": {
            "description": "User object"
          }
        }
      },
      "patch": {
        "summary": "Update current user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "avatarPath": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated user"
          }
        }
      },
      "delete": {
        "summary": "Soft-delete current account",
        "responses": {
          "200": {
            "description": "Acknowledged"
          }
        }
      }
    },
    "/me/password": {
      "post": {
        "summary": "Change password",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "currentPassword": {
                    "type": "string"
                  },
                  "newPassword": {
                    "type": "string"
                  }
                },
                "required": [
                  "currentPassword",
                  "newPassword"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Current password is incorrect"
          }
        }
      }
    },
    "/me/providers": {
      "get": {
        "summary": "Get AI provider keys (masked)",
        "description": "Returns every known provider key name with its value masked to first 6 + last 4 chars, or null when unset. Keys are encrypted at rest in the DB.",
        "responses": {
          "200": {
            "description": "Masked provider keys",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "providers": {
                      "type": "object",
                      "additionalProperties": {
                        "type": [
                          "string",
                          "null"
                        ]
                      }
                    }
                  },
                  "required": [
                    "providers"
                  ]
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update AI provider keys",
        "description": "Partial update. Null value deletes a key; string value sets it. Unknown names return 400.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "providers": {
                    "type": "object",
                    "additionalProperties": {
                      "type": [
                        "string",
                        "null"
                      ]
                    }
                  }
                },
                "required": [
                  "providers"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated masked keys (same shape as GET)"
          },
          "400": {
            "description": "Unknown provider name"
          }
        }
      }
    },
    "/workspaces": {
      "get": {
        "summary": "List workspaces",
        "responses": {
          "200": {
            "description": "Workspace array"
          }
        }
      },
      "post": {
        "summary": "Create workspace",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "icon": {
                    "type": "string"
                  }
                },
                "required": [
                  "name"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created workspace"
          }
        }
      }
    },
...
    "/chats": {
      "get": {
        "summary": "List chats",
        "responses": {
          "200": {
            "description": "Chat array with last-message snippet"
          }
        }
      },
      "post": {
        "summary": "Create chat",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "workspaceId": {
                    "type": "string"
                  },
                  "agentId": {
                    "type": "string"
                  },
                  "title": {
                    "type": "string"
                  },
                  "goal": {
                    "type": "string"
                  }
                },
                "required": [
                  "workspaceId",
                  "agentId",
                  "title"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created chat"
          }
        }
      }
    },
...
    "/library": {
      "get": {
        "summary": "List library files",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Library listing"
          }
        }
      },
      "post": {
        "summary": "Upload to library",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "mime": {
                    "type": "string"
                  },
                  "contentBase64": {
                    "type": "string"
                  }
                },
                "required": [
                  "name",
                  "mime",
                  "contentBase64"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created file"
          }
        }
      }
    },
...
    "/scheduled-jobs": {
      "get": {
        "summary": "List active scheduled jobs",
        "responses": {
          "200": {
            "description": "Array of active scheduled jobs"
          }
        }
      },
      "post": {
        "summary": "Create scheduled job",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "chatId": {
                    "type": "string"
                  },
                  "prompt": {
                    "type": "string"
                  },
                  "mode": {
                    "type": "string",
                    "enum": [
                      "scheduled",
                      "recurring"
                    ]
                  },
                  "spec": {
                    "type": "string"
                  }
                },
                "required": [
                  "prompt",
                  "mode",
                  "spec"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created run"
          }
        }
      }
    },
...
    "/tools/models": {
      "get": {
        "summary": "List AI models that are ready to use",
        "description": "Returns the set of models reachable with currently configured provider credentials (server-wide, not agent-scoped). Every returned model is ready — opencode only surfaces models for providers whose API key is present in the sandbox env. Foundation of host-initiated sandboxed tool calling (ARCHITECTURE.md §7).",
        "parameters": [
          {
            "name": "provider",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Restrict to a single provider id, e.g. \"anthropic\"."
          }
        ],
        "responses": {
          "200": {
            "description": "Array of ready-to-use models",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "description": "Opencode canonical id, e.g. \"anthropic/claude-opus-4-7\"."
                      },
                      "provider": {
                        "type": "string",
                        "description": "Provider portion of id, e.g. \"anthropic\"."
                      }
                    },
                    "required": [
                      "id",
                      "provider"
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "Sandbox rejected the listing"
          },
          "404": {
            "description": "No sandbox available"
          }
        }
      }
    },
    "/search": {
      "get": {
        "summary": "Search across artifacts, chats, and library",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scope",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "all",
                "artifacts",
                "chats",
                "library"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search result array"
          }
        }
      }
    },
    "/ws": {
      "get": {
        "summary": "WebSocket upgrade",
        "description": "Upgrades to WebSocket. Authenticate via ?token= query parameter.",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "101": {
            "description": "Switching Protocols"
          }
        }
      }
    },
...
  }
}